requirements.txt
adal==1.2.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.3.0
botbuilder-core==4.12.0
botbuilder-integration-aiohttp==4.12.0
botbuilder-schema==4.12.0
botframework-connector==4.12.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.3.2
idna==2.10
isodate==0.6.0
jsonpickle==1.2
msal==1.6.0
msrest==0.6.10
multidict==4.7.6
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.5.3
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
tabulate==0.8.9
urllib3==1.25.11
webdriver-manager==3.3.0
yarl==1.4.2
